feat : user 추가, book 추가

+++ src/main/java/com/sparta/library/dto/RentRequestDto.java
@@ -4,8 +4,8 @@ import lombok.Getter;
 
 @Getter
 public class RentRequestDto {
-    private Long user_id;
-    private Long book_id;
+    private int user_id;
+    private int book_id;
     private String status;
     private String rent_date;
     private String return_date;
